Faysal

See merge request disystem/biva-mart!712

+++ app/Repositories/DeliverymanManageRepository.php
@@ -7,6 +7,7 @@ use App\Interfaces\DeliverymanManageInterface;
 use App\Models\DeliveryMan;
 use App\Models\Order;
 use App\Models\OrderDeliveryHistory;
+use App\Models\OrderRefundReason;
 use App\Models\SystemCommission;
 use App\Models\Translation;
 use App\Models\User;
@@ -363,11 +364,15 @@ class DeliverymanManageRepository implements DeliverymanManageInterface
 
     public function deleteVehicle(int $id)
     {
-        try {
-            $vehicle = VehicleType::findorfail($id);
+        $vehicle = VehicleType::find($id);
+        if ($vehicle) {
+            // Delete related translations
+            $this->translation->where('translatable_type', VehicleType::class)
+                ->where('translatable_id', $id)
+                ->delete();
             $vehicle->delete();
             return true;
-        } catch (\Throwable $th) {
+        } else {
             return false;
         }
     }
